Adapt spark to take a RunFunctionRequest on stdin
Previously spark simply passed its stdin to the OCI runtime it wrapped,
and returned stdout from said runtime. It now takes a protobuf
serialized RunFunctionRequest on stdin and returns a RunFunctionResponse
on stdout.

This allows spark to be passed the extra context it needs in order to
run a function - e.g. timeouts, image pull policy, network policy, etc.
Spark 'unwraps' the RunFunctionRequest, passing only the FunctionIO
bytes to the OCI runtime stdin, then wraps the returned FunctionIO in a
RunFunctionResponse.

Signed-off-by: Nic Cope <nicc@rk0n.org>