Rework OCI store to support ImagePullPolicy
This turned out to be a lot more work than expected. The key differences
are that we now:

* Cache a mapping of known OCI references to their image digests.
* Cache all OCI images separately from the OCI bundle stores.

Notably the 'uncompressed' bundler now just makes bundles - it's backed
by layers and config stored in the new OCI image store. The 'overlay'
bundler is also backed by said store, but creates another cache in the
form of extracted overlayfs compatible layer directories. This means
this change roughly doubles the cache disk space used in overlayfs mode.

Signed-off-by: Nic Cope <nicc@rk0n.org>