Add a Composer that uses both P&T and Composition Functions
The idea here is that we completely swap out the Composition
implementation when Composition Functions are enabled. I had considered
implementing Functions as a second stage after P&T, but decided it
wouldn't work well. The main issue is that we'd have to persist
everything from P&T Composition to the APIs server then fun functions,
which could cause flapping/fighting.

Instead the new flow is roughly:

1. Get existing state (composed resources).
2. Build FunctionIO observed object.
3. Do P&T Composition.
4. Build initial FunctionIO desired object (from P&T output).
5. Do Function Composition.
6. Persist all the desired state.
7. Derive XR readiness, connection details, etc.

Signed-off-by: Nic Cope <nicc@rk0n.org>