Build crun, instead of curling it
This takes forever and is generally worse in every way, except that I
expect it to build on all our supported Linux architectures.

I may need to give up and use runc, which releases binaries for all of
the architectures we support. That said, it seems ideal that we wouldn't
be curl-ing dependencies on each build.

Signed-off-by: Nic Cope <nicc@rk0n.org>