Use GVK+Name for object reference sort order on package revisions
Sorting of object references became necessary after introducing
concurrent establishment of package resources because the reference
order was no longer deterministic. To fix the issue, sorting references
based on UUID was introduced. However, inactive package revisions do not
include object UUIDs in their references, causing sorting based on UUID
to be unpredictable.

This updates sort to be based on Group, Version, Kind, and Name, which
is guaranteed to be unique in a cluster. It also has the nice property
of ensuring that the object references preserve the same order any time
a given package is installed in any cluster, which is not an attribue of
UUID-based sort.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>