Use unstable sort in package revision object references
Switches to using unstable sort when sort object references for a
package revision. Stable sort does provider any additional value in this
case as we are not concerned with preserving existing order in the
slice, which is non-deterministic across reconciles, but rather in the
status of the package revision. We also should not have duplicate
resources in a single package, but if we do the sort order of equivalent
references would not matter as the references would be identical (i.e.
switching their order would not cause a diff in the package revision
status).

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>
(cherry picked from commit 6411a86265cefc3095a48e49e638976ba3fa9832)