Just call them "connection secret keys"
I wanted to move to saying "connection detail keys" in FunctionIO to
differentiate from specifically Kubernetes secrets. This is confusing
though, because the old "connection secret key" exists in a lot of
places. It's not strictly wrong; connection details come from a "secret
store" so let's just leave it as is.

Signed-off-by: Nic Cope <nicc@rk0n.org>