Fix 'slice bounds out of range' crash in NewCompositionRevision
Previously we took a string of arbitrary length then tried to slice it.
In practice the caller should have always passed a sufficiently long
string, but in theory they didn't have to.

Now we just create the hash inside the function. This is a little slower
because we now hash twice (in the caller and in NewCompositionRevision),
but it is safer.

Signed-off-by: Nic Cope <nicc@rk0n.org>