feat: ClampMin and ClampMax math transforms Usecase:
For building cloud-agnostic bucket lifecycle policies, I have a parameter called `expireObjectsDays`. I use the value `0` of that parameter to disable object expiry. In AWS, I use a `match` transform to set the `status` of the lifecycle policy to `Disabled` when the days parameter is zero: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/s3_bucket_lifecycle_configuration#status

However, I also have a patch on the actual days attribute. Right now, I haven't found an easy way of setting a minimum of 1 while also keeping the original value if it was already higher. It would be easily done with a clamp:

```yaml
                    - type: FromCompositeFieldPath
                      fromFieldPath: spec.parameters.expireObjectsDays
                      toFieldPath: spec.forProvider.rule[0].expiration[0].days
                      transforms:
                        - type: math
                          math:
                            type: ClampMin
                            clampMin: 1
```

I called the transforms `clampMin` and `clampMax` because it's clearer what it does rather than `min` and `max` (does it clamp to that value OR find the max between the input and given param?)

Signed-off-by: Julien Duchesne <julien.duchesne@grafana.com>