feat: Composition logical validation webhook
This patch introduces a Validating Webhook for Compositions that can be
enabled by setting `webhooks.enabled` to `true` in the provided Helm
chart, alongside the one already available for Composite Resource
Definitions.

This is going to be the first of a series of patches to achieve what has
been proposed in the dedicated design document (PR #3756) and it's
indeed implementing what is there referenced as "logical validation".

With this, upon update or creation, if the webhook is enabled, all
Composition will be checked and rejected in case of any inconsistency
that was not already encoded in the Composition's schema itself, but
that otherwise would have resulted in a runtime error only once selected
for a Composite resource, e.g. Patches not defining required fields for
their type or defining regexes not compiling properly.

In following patches we will address the other types of validations
defined in the design document, so taking into consideration available
schemas.

Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>
Co-authored-by: Lovro Sviben <lovro.sviben@upbound.io>
Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>