feat: In match transform, allow fallback to input This would be useful to make conditional transformations to an input but, if nothing matches, keep the input unmodified As it is, anything that goes through this transform is squashed
Usecase:

One case this is useful is for the `newer_noncurrent_versions` attribute: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/s3_bucket_lifecycle_configuration#newer_noncurrent_versions. It actually takes in a string rather than a number. When the attribute should not be set, it has to be an empty string, not `"0"`, which fails because the value has to be higher than 0

So with this new match feature, I can transform my composition parameter to be replaced by an empty string when it's zero, but keep all other values:
```
                   -  type: FromCompositeFieldPath
                      fromFieldPath: spec.parameters.keepNonCurrentVersions
                      toFieldPath: spec.forProvider.rule[0].noncurrentVersionExpiration[0].newerNoncurrentVersions
                      transforms:
                        - parameters:
                            toType: string
                          type: convert
                        - fallbackTo: Input
                          patterns:
                            - literal: "0"
                              result: ""
                              type: literal
                          type: match

```

Signed-off-by: Julien Duchesne <julien.duchesne@grafana.com>