Use CompositionRevisions in controller code
Previously all of our controller code operated on the Composition type.
When CompositionRevisions were enabled we'd convert them back to a
Composition as part of the Composition-fetching process. This was
because we wanted our code paths to be the same whether
CompositionRevisions were enabled or not.

Now that CompositionRevisions are GA they can't be disabled, and it's
safe for our controller code to use them directly.

Signed-off-by: Nic Cope <nicc@rk0n.org>