Only allow providers in the same OCI 'org' to share a family
We define an org to be the OCI registry and the part before the '/' in
the repository. (As far as I can tell there's no specific name for this
segment).

Requiring providers to be in the same org to be part of the same family
prevents a malicious or misconfigured provider from joining a family
that it should not be a part of. This would be bad, as the malicious
provider would be granted RBAC access to the types in its family.

Signed-off-by: Nic Cope <nicc@rk0n.org>
(cherry picked from commit f1daf31c97b074ebe07ae7f1850267699e4c802f)