Add labels from Crossplane metadata, don't set them
Adding them is a no-op once they've been added, and thus doesn't trigger
an update/reconcile due to metadata changing. When we were setting them
the package manager and package revision reconcilers would fight, with
the package manager trying to add the pkg.crossplane.io/provider label,
and the reconciler package trying to remove it.

Signed-off-by: Nic Cope <nicc@rk0n.org>