Avoid listing CRDs in the RBAC manager
We found in https://github.com/crossplane/crossplane/issues/4016 that
working with CRDs was quite expensive. It turns out we don't need to. In
order to create RBAC roles we only need the plural and group of the
types defined by a ProviderRevision (i.e. the types defined by CRDs
installed by a ProviderRevision). We can derive this from the status of
the ProviderRevision, which has references to all CRDs it creates.

My testing shows that this significantly reduces CPU and memory consumed
by the RBAC manager when installing many (~70) providers from the same
family at once.

Signed-off-by: Nic Cope <nicc@rk0n.org>