Drop CompositionRevision v1alpha1
We promoted this API to v1 in Crossplane v1.12. We're keeping v1beta1
around for a few releases to give folks a chance to migrate tot the v1
API but I think v1alpha1 can go. I actually intended to remove it with
v1.12, but missed deleting this file so it has stuck around.

Signed-off-by: Nic Cope <nicc@rk0n.org>