Implement support for v1beta1 Composition Functions
https://github.com/crossplane/crossplane/blob/ecac16b53ef45d9ddd5a8e737722d31b7bf8ad39/design/design-doc-composition-functions.md

This commit updates Crossplane Composition to support the new v1beta
design of Composition Functions. It contains a breaking change to the
Composition type, in that it removes the 'functions' array and replaces
it with a 'pipeline' array.

This commit also makes some changes to the PTComposer (i.e. the
composition logic that is used when Functions are disabled). This was
necessary to break up the Renderer logic (that is shared by both
composer implementations) into the parts that should run before the
Composition Function pipeline and the parts that should run after it.

Signed-off-by: Nic Cope <nicc@rk0n.org>