fix xfn e2e tests on osx
when docker engine runs inside a VM, it is not possible to
reach k8s nodes from the hosts.

In order to make possible to copy images from host
to the container registry running inside kind cluster,
we need to add extra port mappings to the kind configuration
as suggested in https://kind.sigs.k8s.io/docs/user/configuration/#extra-port-mappings

We start kind with such configuration and later on
when image copying is needed, we detect if our cluster is kind
and use proper address to reach the registry. Otherwise, we fall
back on detecting node address.

Signed-off-by: Predrag Knezevic <predrag.knezevic@upbound.io>