Rename function 'config' to 'input'
This better frames it as what is is - input to a particular Function
call. The downside with calling if 'config' was that folks could mistake
it to mean it was configuring the Function _globally_, not just
configuring how one particular entry in one Composition's pipeline of
steps calls a Function.

Signed-off-by: Nic Cope <nicc@rk0n.org>