avoid hard-coding registry endpoint for image copying
Introduced funcs.ServiceIngressEndPoint(...) that return the endpoint that
can be used for accessing a service deployed in the cluster. If the tests
are running against kind cluster, it consults the kind config and if
particular port forwarding exists for given node port, it returns
the endpoint accessible from host interfaces. Otherwise, it
tries to detect node ip and use the obtained address and the nodeport
to access the service.

Signed-off-by: Predrag Knezevic <predrag.knezevic@upbound.io>