Move muvaf from maintainers to emeritus
It has been an honor to serve as maintainer in Crossplane
project for the last 4 years. My focus has slightly
shifted from infrastructure and I'd like to give room
for the next generation of folks stepping up. I have
full trust that they will be the best stewards of the
project going forward - it's Crossplane, it always
reconciles towards the best desired state.

"And now, I think I am quite ready to go on another
journey. Are you coming?" - Bilbo, the Oldest Hobbit

Signed-off-by: Muvaffak Onus <me@muvaf.com>