Use server-side apply when in mode: Pipeline
This commit updates the FunctionComposer to use SSA to create and update
composed resources, and to update the XR. I don't believe adopting SSA
in this context warrants its own feature flag, since Functions are a new
feature in themselves, and we only use SSA to power Functions.

It also limits Composition Functions to effecting only an XR's status.
The Composition Function pipeline cannot effect an XR's metadata or
spec. This is in service of #4518, which proposes that:

* metadata and spec flow claim -> XR -> composed resources
* status flow composed resources -> XR -> claim

A Function that returns desired XR metadata or spec will be ignored.

Signed-off-by: Nic Cope <nicc@rk0n.org>