Hardwire the XR Reconciler to use *composite.Unstructured
Previously we mostly accepted the resource.Composite interface, but in
practice neither Composer that the Reconciler can use would work with
any underlying type apart from *composite.Unstructured. Having the
Composer interface require that type makes this explicit.

The PTComposer needs an unstructured type in order to fieldpath.Pave
and patch it. The FunctionComposer needs and unstructured type in order
to use server-side apply.

Signed-off-by: Nic Cope <nicc@rk0n.org>