Support Composition environment via context
This commit adds a new field - context - to both RunFunctionRequest and
RunFunctionResponse. Context is an arbitary set of key/value pairs that
broadly represents the "context" in which the Function is running.

Context isn't desired state - Crossplane discards whatever context the
last Function in the pipeline returns. It's also not strictly observed
state, but it could be used to supply ancillary observed state. Consider
the Composition environment, or observed resources that are useful to
the Composite Resource but not _part of it_. You could also imagine
context being used to supply credentials to a Function.

Context is "bootstrapped" by Crossplane, then passed through the
Function pipeline, like desired state, from RunFunctionRequest to
RunFunctionResponse. This means context can originate either from
Crossplane, or from a Function in the pipline.

I see context as having two key uses:

* Allowing Crossplane to inject things into the Function pipeline when
  we're not quite sure of the API (e.g EnvironmentConfig). In this way
  it's a bit like metadata.annotations. We can use it to pass data to
  the Function pipeline without having to actually update the protobuf
  types to accomodate alpha fields we may later make breaking changes
  to.

* Allowing Functions to share context with each other. Currently the
  only thing one Function can pass to another is desired state.

Signed-off-by: Nic Cope <nicc@rk0n.org>