Add ability to pause reconciliation of Provider/Configuration/Function instances
The reconciliation can be paused by adding `crossplane.io/paused: "true"`
to the requested object. Removing the annotation resumes its reconciliation.

As with claims and composites, when paused, object removal are not possible.
We do not explicitly return errors to callers, instead the delete requests
are stuck until the pause annotation is removed.

In order to improve UX, we add to the pause status condition a human readable
message explain the consequences, and in case of deletion we generate
an additional k8s event and a log entry on info level.

Signed-off-by: Predrag Knezevic <predrag.knezevic@upbound.io>