Update top-level help for all commands
This touches only the help:"" struct tags and the Help methods that add
more detailed help for commands. I'm not touching argument or flag help
for now.

In general I've tried to keep these as lightweight as possible. My
thinking is that it's easier to update the docs than to update these
help strings, so we should bias for most information living in the docs.

Signed-off-by: Nic Cope <nicc@rk0n.org>