Try to say "OCI image" less
Frustratingly, there's no clear name for the URL-like thing that
represents an OCI artifact in a registry.

Sometimes it's a "name" or "image name". This is confusing because our
kind: Provider resource also has a "name" - its metadata.name. I've
biased for calling it "a package" and clarifying what it's made up of -
i.e. a repo, tag, and registry.

Signed-off-by: Nic Cope <nicc@rk0n.org>