Add a one-pager guiding Function build tooling
I think as we build templates for languages like Go, Python, and
TypeScript it will be useful to have a framework/philosophy for our
patterns and tooling choices. The ultimate goal here is to make building
a Function in a language you already know and work with regularly a
familiar, unintimidating experience.

Signed-off-by: Nic Cope <nicc@rk0n.org>