Propagate package annotations from config file labels to layer annotations
It's the layer annotations that the package manager actually cares
about, not the labels in the config file. I don't 100% follow, but it
appears that we use the config file as temporary storage for these
annotations because they're not persisted to the tarball format. Or
possibly the tarball format just loses them in certain situations?

Signed-off-by: Nic Cope <nicc@rk0n.org>