Have the claim controller support both CSA and SSA
This reintroduces the old client-side apply claim logic. Instead of two
'Configurators' (one for the XR and one for the claim) it merges the
logic into one 'Syncer'.

It keeps the new server-side apply claim logic, but also merges it from
two 'Configurators' into one 'Syncer'.

This allows us to toggle server-side apply on and off by switching which
Syncer the controller uses, allowing SSA to be put behind a feature
flag.

Signed-off-by: Nic Cope <nicc@rk0n.org>