Have definition reconciler use WithOptions
This was removed in https://github.com/crossplane/crossplane/pull/4637,
with the options instead passed as a 'first-class' argument.

I don't think these need to be a first-class argument. This reverts to
the old behaviour, which matches the claim 'offered' controller.

I noticed this asymmetry while reminding myself how feature flags are
plumbed down to the offered and definition controllers. The former is
responsible for managing claim controllers, the latter XR controllers.

Signed-off-by: Nic Cope <nicc@rk0n.org>