Cap XR requeue backoff to 30 seconds
This will allow us to detect ready XRs slightly faster.

Previously backoff was from 1 to 60 seconds. This means an XR that
wasn't ready in the first 63 seconds would be polled every 60 seconds
until it became ready.

Now backoff is from 1 to 30 seconds. This means an XR that isn't ready
in the first 31 seconds will be polled every 30 seconds until it becomes
ready.

Note that this change affects XRs that are persistently returning
errors, not just unready XRs. The XR reconciler only returns errors
when it can't get the XR or can't update the status of the XR.

Signed-off-by: Nic Cope <nicc@rk0n.org>