Enable all golangci-linters
It seems like new linters are being added with each new version of
golangci-lint. I like the idea of opting out of new linters we don't
like, rather than opting into new linters.

I think there will be several that we'll want to opt out of, which I'll
do in a following commit.

Signed-off-by: Nic Cope <nicc@rk0n.org>