test: Fix TestComposedTemplateGetBaseObject to check for correct type
This test started to fail once a DeepCopy method was added to the
composed.Unstructured type to fix https://github.com/crossplane/crossplane/issues/4970.
Now that the full type is returned via that DeepCopy method, this
unit test needs to check for this full type instead of the type it embeds.

Signed-off-by: Jared Watts <jbw976@gmail.com>