USE GITHUB_REF, not EARTHLY_GIT_BRANCH for S3 uploads
${GITHUB_REF##*/} somewhat confusingly resolves to 'merge' for PRs
and (e.g.) 'master' for merges to a branch.

I'm not sure 'merge' is what we _really_ want (as opposed to the PR
number) but it's what the build submodule does, so keep compatibility
with that.

EARTHLY_GIT_BRANCH was working for master, but was resolving to HEAD for
PR builds.

Signed-off-by: Nic Cope <nicc@rk0n.org>