fix(Earthfile): Build also the go binaries when running +build
The `+multiplatform-build` step includes `+go-multiplatform-build`, so
similarly the `build` step should include `+go-build`.

When calling the `+go-build` target, the OS/arch that is used is the
native, but in some cases, eg on Docker Desktop on Mac OS X, emulation
is used, thus the $NATIVEPLATFORM and $TARGETPLATFORM variables
incorrectly report linux/amd64. The USERPLATFORM var though always
returns the correct OS/arch of the host that is triggering the build.

When calling the `+build` target locally, the expectation is to produce
a binary for our own system, so that we can use/test it afterwards.

Thus, this commit changes the `+build` target to call `+go-build` but
for platform=$USERPLATFORM.

Co-authored-by: Nic Cope <nicc@rk0n.org>
Signed-off-by: Theo Chatzimichos <tampakrap@gmail.com>