Fall back to v1beta1 if a function doesn't implement v1
We want to promote the v1beta1 function proto to v1. This isn't a
breaking change. There's actually no changes - we just want to signal
that it's reached v1.

Crossplane (the client) must continue to support v1beta1 functions (i.e.
servers) in order to remain compatible with existing functions before
they update to SDKs that use the v1 protobuf.

To do this, Crossplane will first attempt to send a v1 RPC. If the
server responds that v1 is unimplemented, it'll fall back to v1beta1.

The v1beta1 protos are now automatically replicated from v1. So all
changes made to v1 will be propagated to v1beta1. This means Crossplane
can work entirely with v1 types internally. When it needs to fall back
to send a v1beta1 RunFunctionRequest, it converts a v1 request by
serializing it to bytes, then deserializing it into a v1beta1 request.
As far as I can tell this is the best way to convert a Go protobuf
message to an identical message of a different type.

Signed-off-by: Nic Cope <nicc@rk0n.org>