Remove native support for EnvironmentConfigs
This removes all the Crossplane code that pertains to the alpha
EnvironmentConfig feature. Namely:

* Selecting and loading EnvironmentConfig resources
* Patching to and from the in-memory environment using native P&T
* Validating environment related native P&T patches

We don't intend to remove support for the EnvironmentConfig type. For
lack of a better place to put it we'll keep it packaged with core
Crossplane and promote it to v1beta1.

Going forward, we recommend folks use function-environment-configs to
select and load EnvironmentConfigs. This function loads environment
configs into a composition function pipeline context key, so any other
function in the pipeline can read from the environment. See the link
below for examples.

https://github.com/crossplane-contrib/function-environment-configs

Signed-off-by: Nic Cope <nicc@rk0n.org>