Have the resolver controller watch package revisions
We want to enqueue a reconcile of the Lock controller when any package
revision changes, but we weren't.

It wasn't watching function revisions at all. It was watching provider
and configuration revisions, but using the Owns() builder method. This
method enqeueus requests for the controller reference of the watched
object. The lock isn't the controller reference (or even an owner
reference) of any package revision, so we wouldn't be enqueuing
reconciles for it.

Signed-off-by: Nic Cope <nicc@rk0n.org>
(cherry picked from commit 51ead1d2701a4cfcd176d0ad59aa6414eb742ed5)