Support resolving dependencies of Function packages
This fixes the PackageDependencyManager to allow Function packages to
depend on other packages. Previously this wasn't supported because we
never tried to parse a Function package's package metadata when loading
dependencies.

I'm guessing no-one noticed this because no-one tried to make a function
package with dependencies. I think doing so is valid though - e.g. a
function that's hardcoded to compose resources from a specific provider
should depend on that provider.

Signed-off-by: Nic Cope <nicc@rk0n.org>