Support modern XRs in various watches and indexes
...by reworking how you specify whether an XR has the legacy schema.

I didn't love the previous NewXRFn pattern, and it didn't fit most of
the indexes and watches that typically don't call composite.New, but
instead explicitly create a composite.Unstructured and set Unstructured
field to a Kubernetes unstructured.Unstructured resource.

At the same time I was reluctant to pass around a boolean field (e.g. as
an argument to composite.NewReconciler) representing whether types
should operate on a legacy XR or not. This is mostly because I find
boolean fields hard to read. e.g. It's not obvious what 'true' means:

_ := composite.NewReconciler(client, gvk, true)

It's also hard to extend a boolean if there's ever some third schema
option in addition to 'Legacy' and 'Modern'.

So instead I've added a Schema field to the composite.Unstructured
resource with values that are today effectively boolean.

Signed-off-by: Nic Cope <nicc@rk0n.org>