Allow Usages to use resources in another namespace
The quintessential use case for this is:

* An XR in 'infra-namespace' represents a Kube cluster
* An XR in 'team-namespace' represents an app deployed to the cluster
* The cluster shouldn't be deleted while the app still exists

The app XR will probably compose the usage, and it can only create
resources in its own namespace. So it'll create the usage in its
namespace, but set the used resource ref to another namespace.

Cluster scoped usages still only work with cluster scoped resources.
They don't support including a namespace in their by/of refs.

Namespaced usages now support including a namespace in their of ref - so
they can use a resource in a different namespace.

The usage must always be by a resource in the same namespace as the
usage. If the usage doesn't specify a namespace in its of ref it'll be
assumed to be a usage of a resource in the same namespace as the usage.

Signed-off-by: Nic Cope <nicc@rk0n.org>