Add a util for finding usages
We always want to make sure no kind of usage uses a thing. So for
example we don't want to mark a cluster scoped Foo as no longer in use
just because no Usage in apiextensions.crossplane.io uses it - it could
still be used by a ClusterUsage in protection.crossplane.io.

Since the usage controller and webhook both had list implementations I
deduped them into this utility type.

Signed-off-by: Nic Cope <nicc@rk0n.org>