Actually compose a CRD in the cluster-scoped XR E2E
I noticed the CRD we were composing was invalid - missing a schema.

Yet the XR was still becoming ready. I think because the function just
told it the CRD was ready, despite it not actually being created
successfully.

I need to look into this more - there's a case in the FunctionComposer
where we'll ignore apply errors due to resources being invalid, but set
the resources to 'unsynced'. So maybe the XR was ready but not synced?

Signed-off-by: Nic Cope <nicc@rk0n.org>