Don't try to start an informer when stopping a source
Before this commit, calling a StoppableSource's Stop method would
potentially start a new informer.

We stop a source by removing its EventHandler from the relevant
informer. Previously each StoppableSource had a map of all informers
(for some reason - not sure why I did this) despite only needing one. So
they'd call GetInformers to get the informer they wanted to remove the
event handler from.

The problem is that GetInformer will implicitly try to start a new
informer if the informer isn't already started. So to stop a source we'd
first need to start its informer. That's not always possible - e.g. if
Crossplane doesn't have RBAC access to the informer's type, or the CRD
for the informer's type no longer exists.

Now each StoppableSource only contains the informer for the type they're
a source of events for, so they don't need to call GetInformer to stop a
watch.

Signed-off-by: Nic Cope <nicc@rk0n.org>