Reduce race in the watch garbage collector
We don't want the watch garbage collector to stop a watch for a composed
resource prematurely. If it did, the XR controller wouldn't be called to
reconcile the XR when that type of composed resource changed.

Before this commit the watch garbage collector determined what watches
were still needed only by listing from cache. If the cache was stale, it
might've incorrectly determined a watch was no longer needed and stopped
it.

After this commit, the watch garbage collector double checks by making
direct calls to the API server to determine what watches are still
needed before it stops any.

The race still exists, but it's much smaller now. A watch would have to
become needed between when Crossplane lists from the API server and when
it calls StopWatches.

Signed-off-by: Nic Cope <nicc@rk0n.org>