Promote realtime composition to beta, don't poll in XR reconciler
After this commit realtime composition is enabled by default. You can
disable it by setting --enable-realtime-composition=false.

When realtime composition is enabled XRs won't requeue on a poll
interval by default.

XRs will requeue a reconcile when one of the following happens:

* The XR changes
* A composed resource changes
* A new CompositionRevision is created, and the XR should auto-update
* Crossplane's sync interval (default 1 hour) is exceeded
* The last function pipeline run's TTL is exceeded

We consider a function pipeline run's TTL to be the shortest non-zero
TTL of any RunFunctionResult returned by the pipeline. If the pipeline's
TTL is non-zero we requeue after that TTL.

This lets functions control when Crossplane will call them again.

If a function needs to be called periodically it should set a TTL. A
function might need to do this if it needs to periodically refresh its
extra resources (Crossplane doesn't watch them) or read an input from an
external system.

Crossplane also reserves the right to use the TTL for caching. This
isn't implemented, but when it is Crossplane may not call the function
_before_ the TTL.

Today many functions return a TTL of 1 minute. This is because
function-sdk-go and function-sdk-python use 1 minute as the default TTL.
I think we should update the SDKs to use a longer default TTL, since the
current default will effectively be equivalent to polling every minute.

Signed-off-by: Nic Cope <nicc@rk0n.org>