Delete native P&T based Usage E2E test
It's functionally identical (no pun intended) to the variant that uses
functions, but the native version has a nasty bug. When it patches the
MR it removes the MR's finalizer, which causes the XR and MR
controllers to fight. This is bad when realtime compositions are
enabled.

We could potentially fix the native P&T variant (e.g. by patching in the
finalizer) but I don't think it's worth it. Apart from this very
specific scenario where we artificially set a finalizer to block the
delete, the composition implementation shouldn't matter. Given native
P&T is deprecated let's drop it and save some time in E2E runs.

Signed-off-by: Nic Cope <nicc@rk0n.org>