Rewrite image paths using ImageConfig when resolving dependencies
Use the new rewrite field in the ImageConfig to rewrite image paths in the
dependency manager. We still use the original package paths for the dependencies
themselves (they're rewritten in the package reconciler), but we need to rewrite
when resolving version constraints so that we select versions available in the
rewritten registry.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>