Cache function responses
This commit adds a new --enable-function-response-cache alpha feature
flag. When the flag is enabled, Crossplane caches function responses to
the filesystem.

Crossplane only caches responses that specify a non-zero TTL in their
response metadata. It caches responses until the TTL expires.

The cache layout is like:

/cache/xfn/my-function/9a65f94eb41d05c12a377ebede6a90ee149f297d

my-function is the name of the function and the SHA256 hash is the tag
from the request's metadata.

The first 8 bytes of each file in the cache are the response's deadline.
The deadline is a binary encoded int64 Unix timestamp after which the
response expires. Crossplane computes the deadline by taking the time at
which the response was received and adding its TTL. The deadline is
always padded to 8 bytes.

The remaining bytes of the file are the protobuf encoded response.

A garbage collector walks the cache directory once per minute and
removes any expired entries.

I chose to cache to the filesystem because I worry the cache might be too big to
fit in memory, and because it's simpler than taking a dependency on e.g.
Redis.

I assume caching to the filesystem will typically be faster than making
network requests to a function. I've added cache read/write latency
metrics, so we can see whether that's true. You can use an emptyDir
volume with the Memory medium to cache to memory if needed.

This commit sets function request tags for the first time. Previously
they were in the request schema, but Crossplane never set them.

Crossplane computes a request tag by encoding the (untagged) request to
protobuf and hashing the resulting bytes using SHA256. So the tag
represents the content of the request.

Protobuf encoding is only stable for a particular build of Crossplane.
The encoded bytes could change if the protobuf schema changed, or
Crossplane updated to a new version of protobuf.

This means:

* Two requests with the same tag will always be identical
* Two identical requests made by the same Crossplane build will always
  have the same tag.
* Two identical requests made by different Crossplane builds may have
  different tag.

I think this is good enough for caching purposes. If you change to a
different build of Crossplane that encodes requests slightly differently
it'll just invalidate the cache.

Signed-off-by: Nic Cope <nicc@rk0n.org>