Drop cache miss reasons from metrics
In retrospect I'm not sure these are very useful.

In particular DeadlineExpired could be misleading. A cache miss due to
an expired deadline could just as easily have reason NotFound if the
garbage collector deleted the expired response from the cache before
Crossplane sent the request.

EmptyRequestTag and EmptyResponse will probably never happen in
practice. The former would be a programming error in Crossplane (i.e.
the caller not setting the tag) and the latter would only happen if the
cache was corrupted or tampered with such that a cached file was exactly
8 bytes.

Error is probably the most useful for metrics, but you could probably
spot cache misses due to errors in a graph of the two counts pretty
easily.

Signed-off-by: Nic Cope <nicc@rk0n.org>