Encode cached responses as (only) protobuf messages
Previously they were the response encoded as a protobuf, prefixed with
the expiry deadline as a little endian encoded uint64 Unix timestamp.

This switches to using pure protobufs. It adds a new internal protobuf
message used only for the cache. I chose to make the new message
internal because it's a Crossplane implementation detail - not intended
for data interchange with other systems.

This approach required some changes to our code generation pipeline,
which previously assumed all generated code lived under apis/. I like
the simpler code though.

One downside compared to the previous approach is that we'll now read
the entire message into memory even if the deadline has expired - e.g.
in the garbage collector. Previously we'd only read the deadline for
expired responses.

Signed-off-by: Nic Cope <nicc@rk0n.org>