Factor revision activation out of the main package reconciler
In preparation for supporting multiple active function revisions, move the
revision activation logic from the package reconciler into a new injectable
dependency, the RevisionActivator. For now, the behavior hasn't changed: when
the activation policy is manual it does nothing and when the policy is automatic
it makes only the current revision active.

This change also fixes a confusing state with the manual activation
policy. Previously, when a new revision was created *all* revisions would be
marked inactive, making it hard to tell which revision's resources and runtime
were installed. Now, we never update the activation state of revisions for
packages using the manual policy: that's up to the user.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>