composite: Add function revision selectors to pipeline steps
Add two new fields, functionRevisionRef and functionRevisionSelector, to
pipeline steps, allowing a function revision other than the active one to be
used.

When the function revision selector feature flag is enabled, use these new
fields to look up function revisions in the function runner. When the feature
flag is disabled, use the existing logic to always select the active revision.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>