Tune XR rate limiter to allow more retries at first
A new XR needs a couple of reconciles in a row to do its thing. If we
back off from 1s -> 30s we hit the 30s cap in less than 10 reconciles.

Now we'll queue at 2, 4, 8, 16, 32ms etc before we cap at 30 seconds.

Ideally we'd use a rate limiter that allowed a burst before backing off.

Signed-off-by: Nic Cope <nicc@rk0n.org>