Disable wsl_v5 linter
I think this is the linter that's been adding whitespace everywhere
throughout the codebase. I thought it was gofumpt since it's a form of
formatting and I couldn't figure out which of golangci-lint's formatters
were doing this.

I think it's actually a linter that is auto "fixing" the lack of
whitespace (aka formatting)...

Either way I don't think we really care about the opinions of this
somewhat obscure linter, and we'd already tried to disable it. At this
point I merged a PR that applied its whitespace opinions to the existing
codebase but I don't really want local formatters to fight it going
forward.

Signed-off-by: Nic Cope <nicc@rk0n.org>