feat: add xpkg batch command
This command is used by family providers to batch build their smaller
providers in parallel batches. It makes uses of the existing xpkg push
functionality to push the images to the remote.

This implementation is largely ported from up xpkg batch with minor
changes for readability and maintainability, as well as removing
Upbound specific functionality.  The core functionality should be
largely untouched, which should help minimize regressions from a
code path that has been exercised a fair amount over the years.

Signed-off-by: Jared Watts <jbw976@gmail.com>