Don't watch FunctionRevisions in the Operation controller
Operations are dependent on FunctionRevisions. They use them to check
function capabilities, and to figure out where to call the function.

I was watching FunctionRevisions so we'd queue a reconcile as soon as
the FunctionRevision controller populated a revision's capabilities. The
problem is this can cause a few rapid (un-rate-limited) reconciles of
the Operation, causing it to hit its retry limit very quickly.

So instead we don't watch FunctionRevisions. We just return an error,
and the next retry is queued with exponential backoff.

Signed-off-by: Nic Cope <nicc@rk0n.org>