Fix CRD-to-MRD converter to preserve provider configuration CRDs
The CRD-to-MRD converter was converting all CRDs to MRDs, including provider
configuration types like ProviderConfig, ClusterProviderConfig, and
ProviderConfigUsage. These are not managed resources and should remain as
regular CRDs that get installed immediately.

This updates the converter to use isManagedResource() to identify which CRDs
represent actual managed resources. Provider configuration types are excluded
from conversion and remain as regular CRDs in the output. The converter function
is renamed from mrdObjs to converted and documentation is updated to clarify that
the output contains a mix of MRDs and unchanged CRDs.

Signed-off-by: Nic Cope <nicc@rk0n.org>
(cherry picked from commit d10d85a94b8fd94aa07420265916fe82a633da4c)