fix(xrd): Use per-field validation rules
Replace property-wide self == oldSef validation rule with validation of
individual properties to allow chainging the value of properties that do
not affect the naming or behaviour of CRDs/XRDs, like categories and
shortnames.

Also, remove the validation rule for enforcedCompositionRef as this
should be able to be changed in the future if, say, the ops teams
decides to enforce a newer composition. With the rule enabled it would
be impossible to change the enforced composition ever again without
deleting the whole XRD and risk deleting all underlying claims and
composite resources.

Signed-off-by: MisterMX <mbxd12@web.de>